Fixed LOC reporting for non-UTF-8 files (#1047)
Fixes: https://github.com/joernio/joern/issues/1045

+++ joern-cli/frontends/c2cpg/src/main/scala/io/joern/c2cpg/utils/Report.scala
@@ -12,7 +12,7 @@ object Report {
 
   private type Reports = TrieMap[FileName, ReportEntry]
 
-  private case class ReportEntry(loc: Long, parsed: Boolean, cpgGen: Boolean, duration: Long) {
+  private case class ReportEntry(loc: Int, parsed: Boolean, cpgGen: Boolean, duration: Long) {
     def toSeq: Seq[String] = {
       val lines     = loc.toString
       val dur       = if (duration == 0) "-" else TimeUtils.pretty(duration)
@@ -80,7 +80,7 @@ class Report {
 
   def addReportInfo(
     fileName: FileName,
-    loc: Long,
+    loc: Int,
     parsed: Boolean = false,
     cpgGen: Boolean = false,
     duration: Long = 0
